labwork3/dummyAPI/requirements.txt
Flask 
request